backend/requirements.txt
# AI Native 研发平台 - Python Dependencies

# Flask
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Database
PyMySQL==1.1.0
cryptography==41.0.7

# Scheduler
APScheduler==3.10.4

# HTTP
requests>=2.32.3

# YAML
PyYAML==6.0.1

# Marshmallow
marshmallow==3.20.1

# Environment
python-dotenv==1.0.0

# JWT
PyJWT==2.8.0
